chore: reorganise client code
