+++ packages/repack/client.js
@@ -1,1 +1,1 @@
-export * from './dist/client/chunks-api';
+export * from './dist/client/api';
